refactor(site-header-standardised): specs file, test and story adapted - TWIG-151
