+++ src/ec/packages/ec-component-site-header-standardised/demo/data--en.js
@@ -1,0 +1,4 @@
+import specs from '@ecl/ec-specs-site-header/demo/data--en';
+import adapter from '../adapter';
+
+export default adapter(specs);
