+++ src/ec/packages/ec-component-site-header-standardised/site-header-standardised.story.js
@@ -1,3 +1,4 @@
+import merge from 'deepmerge';
 import { storiesOf } from '@storybook/html';
 import { withNotes } from '@ecl-twig/storybook-addon-notes';
 import withCode from '@ecl-twig/storybook-addon-code';
@@ -6,7 +7,8 @@ import defaultSprite from '@ecl/ec-resources-icons/dist/sprites/icons.svg';
 import englishBanner from '@ecl/ec-resources-logo/logo--en.svg';
 import frenchBanner from '@ecl/ec-resources-logo/logo--fr.svg';
 
-import { englishData, frenchData } from './demo/data';
+import englishData from './demo/data--en';
+import frenchData from './demo/data--fr';
 
 import siteHeaderStandardised from './ecl-site-header-standardised.html.twig';
 import notes from './README.md';
@@ -21,29 +23,47 @@ storiesOf('Components/Site Headers/Standardised', module)
   .addDecorator(withCode)
   .add(
     'default',
-    () => {
-      englishData.logged = false;
-      return siteHeaderStandardised(englishData);
-    },
+    () =>
+      siteHeaderStandardised(
+        merge(englishData, {
+          logo: {
+            src: frenchBanner,
+          },
+          icon_file_path: defaultSprite,
+          logged: false,
+        })
+      ),
     {
-      notes: { markdown: notes },
+      notes: { markdown: notes, json: englishData },
     }
   )
   .add(
     'logged in',
-    () => {
-      englishData.logged = true;
-      return siteHeaderStandardised(englishData);
-    },
+    () =>
+      siteHeaderStandardised(
+        merge(englishData, {
+          logo: {
+            src: englishBanner,
+          },
+          icon_file_path: defaultSprite,
+          logged: true,
+        })
+      ),
     {
       notes: { markdown: notes, json: englishData },
     }
   )
   .add(
     'translated',
-    () => {
-      return siteHeaderStandardised(frenchData);
-    },
+    () =>
+      siteHeaderStandardised(
+        merge(frenchData, {
+          logo: {
+            src: frenchBanner,
+          },
+          icon_file_path: defaultSprite,
+        })
+      ),
     {
       notes: { markdown: notes, json: frenchData },
     }
